Replace separate VACUUM ANALYZE with one
"With no parameter, VACUUM processes every table in the current database that the current user has permission to vacuum. With a parameter, VACUUM processes only that table. " 
https://www.postgresql.org/docs/10/sql-vacuum.html

This should help in future if the test database changes structure or tables.  This way we know that all tables are always in testdb are VACUUM'd.

I've tested this on my system by reviewing table statistics on testdb.